Update Case_Study_1_Dataset
In the first case study, we were provided with data in the .xlsx format. Our initial task was to create a database called "job" and insert the data into this newly created database. This involved setting up the necessary structure and schema for the database, as well as importing the data from the .xlsx file into the corresponding tables within the database. By doing so, we were able to organize and store the data in a structured and accessible manner, enabling us to perform further analysis and queries on the dataset. This step was crucial in laying the foundation for subsequent data processing and analysis tasks.